+++ packages/components/CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Added
+- `tabbable` prop to `Link`, `Link Pure` and `Link Social`
+
 ### [2.8.0-beta.2] - 2021-12-22
 
 #### Fixed